+++ storage/connect/value.cpp
@@ -1738,7 +1738,7 @@ DECVAL::DECVAL(PSZ s) : TYPVAL<PSZ>(s)
   if (s) {
     char *p = strchr(Strp, '.');
 
-    Prec = (p) ? Len - (p - Strp) : 0;
+    Prec = (p) ? (int)(Len - (p - Strp)) : 0;
     } // endif s
 
   Type = TYPE_DECIM;
@@ -2647,7 +2647,7 @@ bool DTVAL::SetValue_char(const char *p, int n)
 			// Trim trailing blanks
 			for (p2 = p + n -1; p < p2 && *p2 == ' '; p2--);
 
-			if ((rc = (n = p2 - p + 1) > Len))
+			if ((rc = (n = (int)(p2 - p + 1)) > Len))
 				n = Len;
 
 			memcpy(Sdate, p, n);
